src/requirements.txt
aiokafka==0.8.1
aiven-client==2.19.0
anyio==3.7.0
async-timeout==4.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.99.1
idna==3.4
importlib-metadata==6.7.0
kafka-python==2.0.2
linkify-it-py==2.0.2
markdown-it-py==3.0.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
packaging==23.1
pydantic==1.10.10
Pygments==2.15.1
requests==2.31.0
rich==13.4.2
sniffio==1.3.0
starlette==0.27.0
textual==0.29.0
typing_extensions==4.7.1
uc-micro-py==1.0.2
urllib3==2.0.3
zipp==3.15.0
